+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Versions follows PEP440 version scheme.
 ************
 
 ## Fixed
+* fix: [RGG-935] it's possible to set "End interval" date earlier than "Start interval"
 * fix: [RGG-936] Interval filter breaks after clicking "CLEAR" on empty field
 * fix: [RGG-908] Fix for custom selects in active state